CMSProBaseGateway/src/test/javascript/spec/app/entities/CMSProBaseService/states/states-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';

import AlertService from '@/shared/alert/alert.service';
import * as config from '@/shared/config/config';
import StatesUpdateComponent from '@/entities/CMSProBaseService/states/states-update.vue';
import StatesClass from '@/entities/CMSProBaseService/states/states-update.component';
import StatesService from '@/entities/CMSProBaseService/states/states.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const i18n = config.initI18N(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.component('font-awesome-icon', {});

describe('Component Tests', () => {
  describe('States Management Update Component', () => {
    let wrapper: Wrapper<StatesClass>;
    let comp: StatesClass;
    let statesServiceStub: SinonStubbedInstance<StatesService>;

    beforeEach(() => {
      statesServiceStub = sinon.createStubInstance<StatesService>(StatesService);

      wrapper = shallowMount<StatesClass>(StatesUpdateComponent, {
        store,
        i18n,
        localVue,
        router,
        provide: {
          alertService: () => new AlertService(store),
          statesService: () => statesServiceStub,
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: '123' };
        comp.states = entity;
        statesServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(statesServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.states = entity;
        statesServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(statesServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });
  });
});
